move ajax facet functionality to a facet class
