+++ includes/Init.php
@@ -54,6 +54,13 @@ class Init {
 	public $modules;
 
 	/**
+	 * CP Library facets
+	 *
+	 * @var Facets\Init
+	 */
+	public $facets;
+
+	/**
 	 * Only make one instance of Init
 	 *
 	 * @return Init
@@ -283,6 +290,7 @@ class Init {
 		}
 
 		$this->modules = Modules\Init::get_instance();
+		$this->facets  = Facets\Init::get_instance();
 	}
 
 	/**
